feat: Add configurable pattern whitelisting
This commit refactors the code to include a whitelist for file nesting keys in the configuration. The whitelist is used to filter the patterns for file nesting. If the whitelist is not empty, only the keys included in the whitelist will be returned. This improves the flexibility and customization of the file nesting feature.