Removed PDF from LaTeX files
Compiling a LaTeX document causes a lot of files to be generated.
However, the PDF is the final result, and in my opinion this shall be
not grouped, because you want to open it in VS-Code easily. Therfore, I
removed PDFs from the list. I hope that this change makes sense and does
not affect anybodys work flow.